Moving appropriate localstack container tests
To separate package
