+++ localstack/localstack_container_test.go
@@ -1,12 +1,11 @@
-package localstack
+package localstack_test
 
 import (
 	"context"
-	"fmt"
 	"testing"
 	"unicode/utf8"
 
-	"github.com/callumkerredwards/localstack-go-wrapper/localstack/services"
+	"github.com/callumkerredwards/localstack-go-wrapper/localstack"
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/client"
 	"github.com/stretchr/testify/assert"
@@ -14,7 +13,7 @@ import (
 
 func TestCreatesContainer(t *testing.T) {
 	//when
-	container, err := New()
+	container, err := localstack.New()
 
 	//then
 	assert.NoError(t, err, "New should not throw an error if docker is installed and working")
@@ -25,7 +24,7 @@ func TestCreatesContainer(t *testing.T) {
 
 func TestStartsContainer(t *testing.T) {
 	//given
-	container, err := New()
+	container, err := localstack.New()
 	assert.NoError(t, err)
 
 	//when
@@ -39,7 +38,7 @@ func TestStartsContainer(t *testing.T) {
 
 func TestStopsContainer(t *testing.T) {
 	//given
-	container, err := New()
+	container, err := localstack.New()
 	assert.NoError(t, err)
 	err = container.Start()
 	assert.NoError(t, err)
@@ -50,67 +49,6 @@ func TestStopsContainer(t *testing.T) {
 	//then
 	assert.NoError(t, err, "Stop should stop docker container successfully")
 	assert.NotContains(t, runningDockerContainerIds(t), container.ContainerID)
-}
-
-func TestNoServicesProvidedToContainerConfig(t *testing.T) {
-	//when
-	container, err := New()
-
-	assert.NoError(t, err, "New should not throw an error if docker is installed and working")
-	assert.Len(t, container.containerConfig.Env, 0, "Should not provide any env variables if no services are set")
-}
-
-func TestServicesProvidedToContainerConfig(t *testing.T) {
-	//given
-	sqsConfig := &services.ServiceConfig{
-		Service: services.SQS,
-	}
-	redshiftConfig := &services.ServiceConfig{
-		Service: services.Redshift,
-	}
-	expectedEnv := "SERVICES=sqs,redshift"
-
-	//when
-	container, err := New(sqsConfig, redshiftConfig)
-
-	//then
-	assert.NoError(t, err, "New should not throw an error if docker is installed and working")
-	assert.Len(t, container.containerConfig.Env, 1)
-	assert.Equal(t, expectedEnv, container.containerConfig.Env[0])
-}
-
-func TestNoServicesProvidedToHostConfig(t *testing.T) {
-	//when
-	container, err := New()
-
-	//then
-	assert.NoError(t, err, "New should not throw an error if docker is installed and working")
-	assert.Len(t, container.hostConfig.PortBindings, 17, "Should have mappings for all supported services")
-	m := fmt.Sprintf("%v", container.hostConfig.PortBindings)
-	assert.Contains(t, m, "4567/tcp:[{0.0.0.0 4567}]")
-	assert.Contains(t, m, "4577/tcp:[{0.0.0.0 4577}]")
-	assert.Contains(t, m, "4584/tcp:[{0.0.0.0 4584}]")
-}
-
-func TestServicesProvidedToHostConfig(t *testing.T) {
-	//given
-	sqsConfig := &services.ServiceConfig{
-		Service: services.SQS,
-		Port:    3003,
-	}
-	redshiftConfig := &services.ServiceConfig{
-		Service: services.Redshift,
-	}
-
-	//when
-	container, err := New(sqsConfig, redshiftConfig)
-
-	//then
-	assert.NoError(t, err, "New should not throw an error if docker is installed and working")
-	assert.Len(t, container.hostConfig.PortBindings, 2)
-	m := fmt.Sprintf("%v", container.hostConfig.PortBindings)
-	assert.Contains(t, m, "4576/tcp:[{0.0.0.0 3003}]")
-	assert.Contains(t, m, "4577/tcp:[{0.0.0.0 4577}]")
 }
 
 func runningDockerContainerIds(t *testing.T) []string {
